rename raw vanadium correction with algo sugar
